Move the Spotify Connect code into separate file
