+++ __init__.py
@@ -24,9 +24,7 @@ speaker on the equipment.)
 import re
 from mycroft.skills.core import intent_handler, intent_file_handler
 from mycroft.util.parse import match_one, fuzzy_match
-from mycroft.util.log import LOG
 from mycroft.api import DeviceApi
-from padatious import IntentContainer
 from requests import HTTPError
 from adapt.intent import IntentBuilder
 
@@ -35,8 +33,8 @@ from subprocess import Popen, DEVNULL
 import signal
 from socket import gethostname
 
-import spotipy
-from spotipy.oauth2 import SpotifyClientCredentials
+from .spotify import (MycroftSpotifyCredentials, SpotifyConnect,
+                      get_album_info, get_artist_info, get_song_info)
 import random
 
 from mycroft.skills.common_play_skill import CommonPlaySkill, CPSMatchLevel
@@ -56,228 +54,6 @@ class PlaylistNotFoundError(Exception):
 
 class SpotifyNotAuthorizedError(Exception):
     pass
-
-
-def get_token(dev_cred):
-    """ Get token with a single retry.
-    Args:
-        dev_cred: OAuth Credentials to fetch
-     """
-    retry = False
-    try:
-        d = DeviceApi().get_oauth_token(dev_cred)
-    except HTTPError as e:
-        if e.response.status_code == 404:  # Token doesn't exist
-            raise
-        if e.response.status_code == 401:  # Device isn't paired
-            raise
-        else:
-            retry = True
-    if retry:
-        d = DeviceApi().get_oauth_token(dev_cred)
-    return d
-
-
-class MycroftSpotifyCredentials(SpotifyClientCredentials):
-    """ Credentials object renewing through the Mycroft backend."""
-    def __init__(self, dev_cred):
-        self.dev_cred = dev_cred
-        self.access_token = None
-        self.expiration_time = None
-        self.get_access_token()
-
-    def get_access_token(self):
-        if not self.access_token or time.time() > self.expiration_time:
-            d = get_token(self.dev_cred)
-            self.access_token = d['access_token']
-            # get expiration time from message, if missing assume 1 hour
-            self.expiration_time = d.get('expiration') or time.time() + 3600
-        return self.access_token
-
-
-class SpotifyConnect(spotipy.Spotify):
-    """ Implement the Spotify Connect API.
-    See:  https://developer.spotify.com/web-api/
-
-    This class extends the spotipy.Spotify class with Spotify Connect
-    methods since the Spotipy module including these isn't released yet.
-    """
-    def get_devices(self):
-        """ Get a list of Spotify devices from the API.
-
-        Returns:
-            list of spotify devices connected to the user.
-        """
-        try:
-            # TODO: Cache for a brief time
-            devices = self._get('me/player/devices')['devices']
-            return devices
-        except Exception as e:
-            LOG.error(e)
-
-    def status(self):
-        """ Get current playback status (across the Spotify system) """
-        try:
-            return self._get('me/player/currently-playing')
-        except Exception as e:
-            LOG.error(e)
-            return None
-
-    def is_playing(self, device=None):
-        """ Get playback state, either across Spotify or for given device.
-        Args:
-            device (int): device id to check, if None playback on any device
-                          will be reported.
-
-        Returns:
-            True if specified device is playing
-        """
-        try:
-            status = self.status()
-            if not status['is_playing'] or device is None:
-                return status['is_playing']
-
-            # Verify it is playing on the given device
-            dev = self.get_device(device)
-            return dev and dev['is_active']
-        except:
-            # Technically a 204 return from status() request means 'no track'
-            return False  # assume not playing
-
-    def transfer_playback(self, device_id, force_play=True):
-        """ Transfer playback to another device.
-        Arguments:
-            device_id (int):      transfer playback to this device
-            force_play (boolean): true if playback should start after
-                                  transfer
-        """
-        data = {
-            'device_ids': [device_id],  # Doesn't allow more than one
-            'play': force_play
-        }
-        try:
-            return self._put('me/player', payload=data)
-        except Exception as e:
-            LOG.error(e)
-
-    def play(self, device, uris=None, context_uri=None):
-        """ Start playback of tracks, albums or artist.
-
-        Can play either a list of uris or a context_uri for things like
-        artists and albums. Both uris and context_uri shouldn't be provided
-        at the same time.
-
-        Args:
-            device (int):      device id to start playback on
-            uris (list):       list of track uris to play
-            context_uri (str): Spotify context uri for playing albums or
-                               artists.
-        """
-        data = {}
-        if uris:
-            data['uris'] = uris
-        elif context_uri:
-            data['context_uri'] = context_uri
-        path = 'me/player/play?device_id={}'.format(device)
-        try:
-            self._put(path, payload=data)
-        except Exception as e:
-            LOG.error(e)
-            raise
-
-    def pause(self, device):
-        """ Pause user's playback on device.
-
-        Arguments:
-            device_id: device to pause
-        """
-        LOG.debug('Pausing Spotify playback')
-        try:
-            self._put('me/player/pause?device_id={}'.format(device))
-        except Exception as e:
-            LOG.error(e)
-
-    def next(self, device):
-        """ Skip track.
-
-        Arguments:
-            device_id: device id for playback
-        """
-        LOG.info('This was terrible, let\'s play the next track')
-        try:
-            self._post('me/player/next?device_id={}'.format(device))
-        except Exception as e:
-            LOG.error(e)
-
-    def prev(self, device):
-        """ Move back in playlist.
-
-        Arguments
-            device_id: device target for playback
-        """
-        LOG.debug('That was pretty good, let\'s listen to that again')
-        try:
-            self._post('me/player/previous?device_id={}'.format(device))
-        except Exception as e:
-            LOG.error(e)
-
-    def volume(self, device, volume):
-        """ Set volume of device:
-
-        Parameters:
-            device: device id
-            volume: volume in percent
-        """
-        uri = 'me/player/volume?volume_percent={}&device_id={}'.format(volume,
-                                                                       device)
-        try:
-            self._put(uri)
-        except Exception as e:
-            LOG.error(e)
-
-    def shuffle(self, state):
-        """ Toggle shuffling
-
-            Parameters:
-                state: Shuffle state
-        """
-        uri = 'me/player/shuffle?state={}'.format(state)
-        try:
-            self._put(uri)
-        except Exception as e:
-            LOG.error(e)
-
-
-def get_album_info(data):
-    """ Get album info from data object.
-    Arguments:
-        data: data structure from spotify
-    Returns: tuple with name, [artists], uri)
-    """
-    return (data['albums']['items'][0]['name'],
-            [a['name'] for a in data['albums']['items'][0]['artists']],
-            data['albums']['items'][0]['uri'])
-
-
-def get_artist_info(data):
-    """ Get artist info from data object.
-    Arguments:
-        data: data structure from spotify
-    Returns: tuple with name, uri)
-    """
-    return (data['artists']['items'][0]['name'],
-            data['artists']['items'][0]['uri'])
-
-
-def get_song_info(data):
-    """ Get song info from data object.
-    Arguments:
-        data: data structure from spotify
-    Returns: tuple with name, [artists], uri)
-    """
-    return (data['tracks']['items'][0]['name'],
-            [a['name'] for a in data['tracks']['items'][0]['artists']],
-            data['tracks']['items'][0]['uri'])
 
 
 class SpotifySkill(CommonPlaySkill):
